# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Grok2API
  version: 1.0.0-go
  description: |
    Grok CLI credential gateway. Inference endpoints use OpenAI/Anthropic-shaped
    errors. Admin endpoints use the `{ok,data,error}` envelope.
servers:
  - url: /
...
paths:
...
  /v1/models:
    get:
      tags: [inference]
      summary: List models
      operationId: listModels
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      responses:
        "200":
          description: OpenAI-style model list
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
...
  /v1/chat/completions:
    post:
      tags: [inference]
      summary: Chat Completions
      operationId: createChatCompletion
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Chat completion or SSE stream
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "503":
          $ref: "#/components/responses/OpenAIUnavailable"
  /chat/completions:
    post:
      tags: [inference]
      summary: Chat Completions alias
      operationId: createChatCompletionAlias
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      responses:
        "200":
          description: Chat completion or SSE stream
  /v1/responses:
    post:
      tags: [inference]
      summary: OpenAI Responses
      operationId: createResponse
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Responses payload or SSE stream
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
  /v1/messages:
    post:
      tags: [inference]
      summary: Anthropic Messages
      operationId: createMessage
      security:
        - ClientBearer: []
        - ClientAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Messages payload or SSE stream
        "401":
          $ref: "#/components/responses/OpenAIUnauthorized"
...
  /api/admin/v1/dashboard:
    get:
      tags: [admin]
      summary: Dashboard analytics
      operationId: adminDashboard
      security:
        - AdminBearer: []
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [24h, 7d, 30d]
      responses:
        "200":
          description: Dashboard payload
        "401":
          $ref: "#/components/responses/AdminUnauthorized"
...
  /api/admin/v1/accounts:
    get:
      tags: [admin]
      summary: List accounts
      operationId: adminListAccounts
      security:
        - AdminBearer: []
      parameters:
        - in: query
          name: pool
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Account page
...
  /api/admin/v1/client-keys:
    get:
      tags: [admin]
      summary: List client keys
      operationId: adminListClientKeys
      security:
        - AdminBearer: []
      responses:
        "200":
          description: Key page
    post:
      tags: [admin]
      summary: Create client key
      operationId: adminCreateClientKey
      security:
        - AdminBearer: []
      responses:
        "201":
          description: Created key (secret shown once)
...
  /api/admin/v1/models:
    get:
      tags: [admin]
      summary: List managed models
      operationId: adminListModels
      security:
        - AdminBearer: []
      parameters:
        - in: query
          name: include_disabled
          schema: { type: boolean }
      responses:
        "200":
          description: Model list
...
  /api/admin/v1/settings:
    get:
      tags: [admin]
      summary: Get versioned settings
      operationId: adminGetSettings
      security:
        - AdminBearer: []
      responses:
        "200":
          description: Settings document
    put:
      tags: [admin]
      summary: Update settings (optimistic lock)
      operationId: adminPutSettings
      security:
        - AdminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expected_revision: { type: integer }
                pool: { type: object, additionalProperties: true }
                timeouts: { type: object, additionalProperties: true }
                audit: { type: object, additionalProperties: true }
                proxy: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Updated settings
        "409":
          description: Revision conflict
...
components:
  securitySchemes:
    ClientBearer:
      type: http
      scheme: bearer
      description: Persisted client API key
    ClientAPIKey:
      type: apiKey
      in: header
      name: x-api-key
    AdminBearer:
      type: http
      scheme: bearer
      description: Short-lived administrator access token
...
  responses:
    OpenAIUnauthorized:
      description: Invalid client API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIError"
    OpenAIRateLimited:
      description: Rate limited / quota circuit
      headers:
        Retry-After:
          schema: { type: string }
        X-Grok2API-Pool-Reason:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIError"
    OpenAIUnavailable:
      description: Pool capacity unavailable
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIError"
    AdminUnauthorized:
      description: Administrator authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminEnvelope"
  schemas:
...
    OpenAIError:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: string }
            param: { nullable: true }
    AdminEnvelope:
      type: object
      properties:
        ok: { type: boolean }
        data: { nullable: true }
        error:
          nullable: true
          type: object
          properties:
            code: { type: string }
            message: { type: string }